app/requirements.txt
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.12
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==7.5.0
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
fastavro==1.10.0
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
griffe==1.5.7
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpath-python==1.0.6
logfire-api==3.6.4
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.5.0
mypy-extensions==1.0.0
openai==1.64.0
packaging==24.2
primp==0.14.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-ai==0.0.28
pydantic-ai-slim==0.0.28
pydantic-graph==0.0.28
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
